new rpi things

+++ rpi/autorun.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
-# add your commands here
-echo "\n---executing autorun---"
+# add your commands here, using & to fork (othewise exit will kill them)
 
+enable_autorun=1
 
-# add this to enable reboot if app crashes (put it right after the app launch command) 
-ret=$?
-if [ $ret -ne 0 ]; then
-        sudo reboot
-else
-        echo app exited without errors
+if [ $enable_autorun -eq 1 ]; then
+	echo "-------- executing autorun routine --------"
+
+	cd /home/pi/ofGPIO/apps/synchron/osc_drums
+	./bin/osc_drums
+
+	ret=$?
+
+	if [ $ret -ne 0 ]; then
+        	sudo reboot
+	else
+        	echo app exited without errors
+	fi
 fi
-
 
 exit
